Move the signal handler from `ironfish-rust` to `ironfish-rust-nodejs`
This removes the `libc` dependency from `ironfish-rust`, which is not
supported in generic WASM environments.

In addition to that:

* Removed the `triggerSegfault` function, which does not appear to be
  used anywhere.

* Replaced the call to `libc::exit()` with `std::process::abort()`. The
  two are not exactly equivalent, but the latter is better suited for
  the task. If we want equivalence, we can use `std::process::exit()`
  instead.

* Corrected the definition of `display_trace`: the argument to signal is
  `void (*sighandler_t)(int)`, but our `display_trace` was not accepting
  an int argument. This could lead to stack corruption on some platforms
  (on x86-64, the first argument is passed on a register, so no stack
  corruption can occur there, but other platforms may be affected).

* Changed the `STACK_TRACE` variable from being a `static mut` to a
  stack-local variable. This simplifies the implementation and also
  avoids potential issues with multi-threading.

* Protected the `backtrace_symbols_fd` symbol from being defined on
  platforms where it's not supported.
